wpaicd: report instantaneous signal levels
wpa_supplicant reports stale signal levels for wifi in the BSS
properties, updating them only during an active scan. Since icd2 scans
infrequently (for power management reasons), it therefore reports stale
data most of the time.

Modify icd2 to use the fi.w1.wpa_supplicant1.Interface.SignalPoll()
method for the signal level of the currently active BSS. This method
reports up-to-date values.

Fixes: https://github.com/maemo-leste/bugtracker/issues/730

+++ src/wpaicd.h
@@ -101,7 +101,7 @@ typedef struct {
     gchar *ssid;
     gsize ssid_len;
 
-    gint16 signal;
+    gint32 signal;
 
     gchar *mac_addr;
     gsize mac_addr_len;
